moved routes related to habits into habit.py and fixed other redirects
